onclick function working
Finally managed to get the onClick function to work roughly the way I need it to. Answers once clicked on change the score if correct and question number whether option is correct or incorrect. Still work to do but happy with this ability at the current moment. Tested with cypress and it gives me a green tick